BDA_6_Topic5_Regression_Clustering/requirements.txt
numpy
scipy
scikit-learn
pandas
matplotlib
python-docx
